backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135",
    "uvicorn>=0.42",
    "supabase>=2.28",
    "pydantic[email]>=2.12",
    "pydantic-settings>=2.13",
    "python-multipart>=0.0.22",
    "httpx>=0.28",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "httpx>=0.28",
    "python-dotenv>=1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "strict"
